+++ plugins/woocommerce-admin/packages/components/src/chart/d3chart/utils/line-chart.js
@@ -43,7 +43,7 @@ export const drawLines = ( node, data, params, xOffset ) => {
 			.attr( 'stroke-width', lineStroke )
 			.attr( 'stroke-linejoin', 'round' )
 			.attr( 'stroke-linecap', 'round' )
-			.attr( 'stroke', d => getColor( d.key, params.orderedKeys, params.colorScheme ) )
+			.attr( 'stroke', d => getColor( d.key, params.visibleKeys, params.colorScheme ) )
 			.style( 'opacity', d => {
 				const opacity = d.focus ? 1 : 0.1;
 				return d.visible ? opacity : 0;
@@ -59,7 +59,7 @@ export const drawLines = ( node, data, params, xOffset ) => {
 			.enter()
 			.append( 'circle' )
 			.attr( 'r', dotRadius )
-			.attr( 'fill', d => getColor( d.key, params.orderedKeys, params.colorScheme ) )
+			.attr( 'fill', d => getColor( d.key, params.visibleKeys, params.colorScheme ) )
 			.attr( 'stroke', '#fff' )
 			.attr( 'stroke-width', lineStroke + 1 )
 			.style( 'opacity', d => {
@@ -112,7 +112,7 @@ export const drawLines = ( node, data, params, xOffset ) => {
 		.enter()
 		.append( 'circle' )
 		.attr( 'r', dotRadius + 2 )
-		.attr( 'fill', d => getColor( d.key, params.orderedKeys, params.colorScheme ) )
+		.attr( 'fill', d => getColor( d.key, params.visibleKeys, params.colorScheme ) )
 		.attr( 'stroke', '#fff' )
 		.attr( 'stroke-width', lineStroke + 2 )
 		.attr( 'cx', d => params.xLineScale( moment( d.date ).toDate() ) + xOffset )